+++ admin/js/settings.js
@@ -13,6 +13,8 @@ function valueChanged() {
   document.getElementById("w4os_economy_db_database").parentNode.parentNode.style.display = document.getElementById('w4os_provide_economy_helpers').checked &! document.getElementById('w4os_economy_use_robust_db').checked ? "table-row" : "none";
   document.getElementById("w4os_economy_db_user").parentNode.parentNode.style.display = document.getElementById('w4os_provide_economy_helpers').checked &! document.getElementById('w4os_economy_use_robust_db').checked ? "table-row" : "none";
   document.getElementById("w4os_economy_db_pass").parentNode.parentNode.style.display = document.getElementById('w4os_provide_economy_helpers').checked &! document.getElementById('w4os_economy_use_robust_db').checked ? "table-row" : "none";
+  document.getElementById("w4os_podex_redirect_url").parentNode.parentNode.style.display = document.getElementById('w4os_currency_provider_podex').checked ? "table-row" : "none";
+  document.getElementById("w4os_podex_error_message").parentNode.parentNode.style.display = document.getElementById('w4os_currency_provider_podex').checked ? "table-row" : "none";
 }
 // force check on load
 valueChanged();